+++ ReleaseNotes.md
@@ -22,3 +22,4 @@
 - Add MileGetTickCount function.
 - Add MileCreateThread function.
 - Add MileGetNumberOfHardwareThreads function.
+- Add MileGetWindowsVersion function.